added rpcu and rpcp
Added the lines to create the okcash.conf file and create the rpcu and rpcp so that we don't have to do " ./okcashd " to get them and have to copy them in the okcash.conf. Makes the users save a step and stupid proof about how to use text editors.